src/albibong/requirements.txt
scapy==2.5.0
websockets==12.0
pywebview==5.1
pycryptodome==3.10.1
